requirements.txt
git+https://github.com/deepmind/graphcast.git@main
cartopy
google-cloud-storage
PyYAML
cftime
xarray
numpy
pandas
matplotlib
jax
jaxlib
dm-haiku